refactor all validator modals (create, update, delete) and validator items, cleanup validator store by separating concerns
