+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+
+  "[vue]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+  }
+}
