+++ frontend/package.json
@@ -47,6 +47,7 @@
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.8.0",
+    "@tailwindcss/forms": "^0.5.7",
     "@tsconfig/node20": "^20.1.4",
     "@types/chai": "^4.3.16",
     "@types/jsdom": "^21.1.6",
