+++ frontend/src/views/Simulator/SettingsView.vue
@@ -1,11 +1,13 @@
 <script setup lang="ts">
 import { onMounted } from 'vue'
 import { notify } from '@kyvg/vue3-notification'
-import { TrashIcon } from '@heroicons/vue/24/solid'
 import { useNodeStore } from '@/stores'
-import ConfirmationModal from '@/components/global/ConfirmationModal.vue'
+import ValidatorItem from '@/components/Simulator/ValidatorItem.vue'
+import NewValidatorModal from '@/components/Simulator/NewValidatorModal.vue'
+import { ref } from 'vue'
 
 const nodeStore = useNodeStore()
+const isNewValidatorModalOpen = ref(false)
 
 // Hooks
 onMounted(async () => {
@@ -20,60 +22,6 @@ onMounted(async () => {
     })
   }
 })
-
-async function handleCreateNewValidator() {
-  try {
-    await nodeStore.createNewValidator()
-    notify({
-      title: 'OK',
-      text: 'New validator created',
-      type: 'success',
-    })
-  } catch (error) {
-    console.error(error)
-    notify({
-      title: 'Error',
-      text: (error as Error)?.message || 'Error creating new validator',
-      type: 'error',
-    })
-  }
-}
-
-async function handleUpdateValidator() {
-  try {
-    await nodeStore.updateValidator()
-    notify({
-      title: 'OK',
-      text: 'Validator updated successfully',
-      type: 'success',
-    })
-  } catch (error) {
-    console.error(error)
-    notify({
-      title: 'Error',
-      text: (error as Error)?.message || 'Error udpating the validator',
-      type: 'error',
-    })
-  }
-}
-
-async function handleDeleteValidator() {
-  try {
-    await nodeStore.deleteValidator()
-    notify({
-      title: 'OK',
-      text: 'Validator deleted successfully',
-      type: 'success',
-    })
-  } catch (error) {
-    console.error(error)
-    notify({
-      title: 'Error',
-      text: (error as Error)?.message || 'Error deleting a validator',
-      type: 'error',
-    })
-  }
-}
 
 const handleResetStorage = async () => {
   try {
@@ -95,16 +43,7 @@ const handleResetStorage = async () => {
   }
 }
 
-const handleNumberInput = (event: Event) => {
-  const formattedValue = parseInt((event?.target as any)?.value || '', 10)
 
-  if (!isNaN(formattedValue)) {
-    ;(event?.target as any).value = ''
-    ;(event?.target as any).value = formattedValue
-  } else {
-    event.preventDefault()
-  }
-}
 </script>
 
 <template>
@@ -124,45 +63,15 @@ const handleNumberInput = (event: Event) => {
     </div>
     <div class="flex flex-col" id="tutorial-validators">
       <div class="flex w-full flex-col text-xs">
-        <div
-          data-testid="validator-item-container"
-          class="flex items-center justify-between p-1 px-2 hover:bg-slate-100 dark:hover:bg-zinc-700"
+        <ValidatorItem
           v-for="validator in nodeStore.validators"
           :key="validator.id"
-        >
-          <div
-            class="flex cursor-pointer items-center"
-            data-testid="validator-item"
-            @click="nodeStore.openUpdateValidatorModal(validator)"
-          >
-            <div class="flex text-primary dark:text-white">{{ validator.id }} -</div>
-            <div class="ml-2 flex flex-col items-start">
-              <div class="flex">
-                <span class="mr-1 font-semibold">Model: </span>
-                <span class="text-primary dark:text-white" data-testid="validator-item-model">{{
-                  validator.model
-                }}</span>
-              </div>
-              <div class="flex">
-                <span class="mr-1 font-semibold">Provider: </span>
-                <span data-testid="validator-item-provider">{{ validator.provider }}</span>
-              </div>
-            </div>
-          </div>
-          <div class="flex text-primary dark:text-white">
-            <button
-              @click="nodeStore.openDeleteValidatorModal(validator)"
-              data-testid="validator-item-delete"
-            >
-              <ToolTip text="Delete Validator" :options="{ placement: 'bottom' }" />
-              <TrashIcon class="mr-1 h-4 w-4" />
-            </button>
-          </div>
-        </div>
+          :validator="validator"
+        />
       </div>
     </div>
     <div class="mt-4 flex w-full flex-col px-2">
-      <Btn @click="nodeStore.openCreateNewValidatorModal" data-testid="create-new-validator-btn">
+      <Btn @click="isNewValidatorModalOpen = true" data-testid="create-new-validator-btn">
         New Validator
       </Btn>
     </div>
@@ -182,209 +91,11 @@ const handleNumberInput = (event: Event) => {
         v-if="nodeStore.contractsToDelete.length < 1"
       />
     </div>
-    <Modal :open="nodeStore.createValidatorModalOpen" @close="nodeStore.closeNewValidatorModal">
-      <template #title>Create New Validator</template>
-      <!-- <template #description>Create New Validator</template> -->
 
-      <div class="flex w-full flex-col">
-        <!-- <div class="flex justify-between">
-          <div class="text-xl">Create New Validator</div>
-        </div> -->
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Provider:</p>
-          <select
-            :class="nodeStore.validatorToCreate.provider ? '' : 'border border-red-500'"
-            class="w-full overflow-y-auto bg-slate-100 p-2 dark:bg-zinc-700"
-            data-testid="dropdown-provider-create"
-            v-model="nodeStore.validatorToCreate.provider"
-            required
-          >
-            <option
-              v-for="(_, provider) in nodeStore.nodeProviders"
-              :key="provider"
-              :value="provider"
-              :selected="provider === nodeStore.validatorToCreate.provider"
-            >
-              {{ provider }}
-            </option>
-          </select>
-        </div>
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Model:</p>
-          <select
-            :class="nodeStore.validatorToCreate.model ? '' : 'border border-red-500'"
-            class="w-full overflow-y-auto bg-slate-100 p-2 dark:bg-zinc-700"
-            data-testid="dropdown-model-create"
-            v-model="nodeStore.validatorToCreate.model"
-            required
-          >
-            <option
-              v-for="model in nodeStore.nodeProviders[nodeStore.validatorToCreate.provider]"
-              :key="model"
-              :value="model"
-              :selected="model === nodeStore.validatorToCreate.model"
-            >
-              {{ model }}
-            </option>
-          </select>
-        </div>
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Stake:</p>
-          <input
-            type="number"
-            min="1"
-            step="1"
-            data-testid="input-stake-create"
-            @input="handleNumberInput"
-            v-model="nodeStore.validatorToCreate.stake"
-            :class="nodeStore.validatorToCreate.stake ? '' : 'border border-red-500'"
-            class="w-full bg-slate-100 p-2 dark:bg-zinc-700"
-            required
-          />
-        </div>
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Config:</p>
-
-          <textarea
-            name=""
-            id=""
-            rows="5"
-            cols="60"
-            class="max-h-64 w-full bg-slate-100 p-2 dark:bg-zinc-700"
-            v-model="nodeStore.validatorToCreate.config"
-          >
-          </textarea>
-        </div>
-      </div>
-      <div class="mt-4 flex w-full flex-col">
-        <Btn
-          @click="handleCreateNewValidator"
-          :disabled="!nodeStore.createValidatorModelValid"
-          data-testid="btn-create-validator"
-        >
-          Create
-        </Btn>
-      </div>
-    </Modal>
-    <Modal :open="nodeStore.updateValidatorModalOpen" @close="nodeStore.closeUpdateValidatorModal">
-      <div class="flex flex-col">
-        <div class="flex justify-between">
-          <div class="text-xl">Validator Details</div>
-          <div class="text-primary dark:text-white">ID: {{ nodeStore.selectedValidator?.id }}</div>
-        </div>
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Address:</p>
-
-          <div class="w-full py-2">
-            {{ nodeStore.selectedValidator?.address }}
-          </div>
-        </div>
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Provider:</p>
-          <select
-            :class="nodeStore.validatorToUpdate.provider ? '' : 'border border-red-500'"
-            data-testid="dropdown-provider-update"
-            class="w-full overflow-y-auto bg-slate-100 p-2 dark:bg-zinc-700"
-            v-model="nodeStore.validatorToUpdate.provider"
-            required
-          >
-            <option
-              v-for="(_, provider) in nodeStore.nodeProviders"
-              :key="provider"
-              :value="provider"
-              :selected="provider === nodeStore.validatorToUpdate.provider"
-            >
-              {{ provider }}
-            </option>
-          </select>
-        </div>
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Model:</p>
-          <select
-            :class="nodeStore.validatorToUpdate.model ? '' : 'border border-red-500'"
-            class="w-full overflow-y-auto bg-slate-100 p-2 dark:bg-zinc-700"
-            name="dropdown-model"
-            data-testid="dropdown-model-update"
-            v-model="nodeStore.validatorToUpdate.model"
-            required
-          >
-            <option
-              v-for="model in nodeStore.nodeProviders[nodeStore.validatorToUpdate.provider]"
-              :key="model"
-              :value="model"
-              :selected="model === nodeStore.validatorToUpdate.model"
-            >
-              {{ model }}
-            </option>
-          </select>
-        </div>
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Stake:</p>
-          <input
-            type="number"
-            min="1"
-            step="1"
-            @input="handleNumberInput"
-            v-model="nodeStore.validatorToUpdate.stake"
-            :class="nodeStore.validatorToUpdate.stake ? '' : 'border border-red-500'"
-            data-testid="input-stake-update"
-            class="w-full bg-slate-100 p-2 dark:bg-zinc-700"
-            required
-          />
-        </div>
-        <div class="mt-2 flex flex-col p-2">
-          <p class="text-md font-semibold">Config:</p>
-
-          <textarea
-            name=""
-            id=""
-            rows="5"
-            cols="60"
-            class="max-h-64 w-full bg-slate-100 p-2 dark:bg-zinc-700"
-            v-model="nodeStore.validatorToUpdate.config"
-          >
-          </textarea>
-        </div>
-      </div>
-      <div class="mt-4 flex w-full flex-col">
-        <Btn
-          @click="handleUpdateValidator"
-          :disabled="!nodeStore.updateValidatorModelValid"
-          data-testid="btn-update-validator"
-        >
-          Save
-        </Btn>
-      </div>
-    </Modal>
-
-    <ConfirmationModal
-      :open="nodeStore.deleteValidatorModalOpen"
-      @confirm="handleDeleteValidator"
-      @close="nodeStore.closeDeleteValidatorModal"
-      :buttonText="`Delete Validator #${nodeStore.selectedValidator?.id}`"
-      buttonTestId="btn-delete-validator"
-    >
-      <template #title>Delete Validator</template>
-      <div class="mt-2 flex flex-col p-2">
-        <p class="text-md font-semibold">Address:</p>
-
-        <div class="w-full py-2">
-          {{ nodeStore.selectedValidator?.address }}
-        </div>
-      </div>
-      <div class="mt-2 flex flex-col p-2">
-        <p class="text-md font-semibold">Provider:</p>
-        {{ nodeStore.selectedValidator?.provider }}
-      </div>
-      <div class="mt-2 flex flex-col p-2">
-        <p class="text-md font-semibold">Model:</p>
-        {{ nodeStore.selectedValidator?.model }}
-      </div>
-      <div class="mt-2 flex flex-col p-2">
-        <p class="text-md font-semibold">Stake:</p>
-        {{ nodeStore.selectedValidator?.stake }}
-      </div>
-    </ConfirmationModal>
+    <NewValidatorModal
+      :open="isNewValidatorModalOpen"
+      @close="isNewValidatorModalOpen = false"
+    />
 
     <ConfirmationModal
       :open="nodeStore.resetStorageModalOpen"
